Redo drag and drop to react-beautiful-dnd
