+++ drag-drop-prototype/package.json
@@ -13,6 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "react": "^17.0.2",
+    "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.2",
     "react-draggable": "^4.4.4"
   },
